# billingwebapp/docs/openapi-ai-v1.yaml
openapi: 3.0.3
info:
  title: Clinic AI Integration API
  version: 1.0.0
  description: Secure server-to-server API for the separate Call Assistant. Direct database access is unsupported.
servers:
  - url: /api/v1/ai
security:
  - bearerAuth: []
paths:
...
  /reports/status:
    get:
      summary: Minimum report status for the verified patient
      x-required-scope: report:read
      parameters:
        - {$ref: '#/components/parameters/PatientSession'}
        - {name: report_id, in: query, schema: {type: integer, minimum: 1}}
        - {name: test_name, in: query, schema: {type: string, maxLength: 120}}
        - {name: date_from, in: query, schema: {type: string, format: date}}
        - {name: date_to, in: query, schema: {type: string, format: date}}
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Offset'}
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /reports/{report_id}/secure-link:
    post:
      summary: Generate a short-lived one-time patient-owned report link
      x-required-scope: report:read
      parameters: [{$ref: '#/components/parameters/ReportId'}, {$ref: '#/components/parameters/PatientSession'}]
      responses: {'201': {$ref: '#/components/responses/Success'}, '403': {$ref: '#/components/responses/Forbidden'}}
  /reports/{report_id}/send:
    post:
      summary: Send an approved REPORT_LINK template idempotently
      x-required-scope: report:send
      parameters: [{$ref: '#/components/parameters/ReportId'}, {$ref: '#/components/parameters/PatientSession'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {$ref: '#/components/requestBodies/Channel'}
      responses: {'202': {$ref: '#/components/responses/Success'}, '409': {$ref: '#/components/responses/Conflict'}}
  /documents/{token}:
    get:
      summary: Consume a short-lived one-time document token
      security: []
      parameters: [{name: token, in: path, required: true, schema: {type: string}}]
      responses: {'200': {description: Private report stream}, '403': {$ref: '#/components/responses/Forbidden'}}
...
components:
  securitySchemes:
    basicAuth: {type: http, scheme: basic}
    bearerAuth: {type: http, scheme: bearer, bearerFormat: opaque}
  parameters:
...
    ReportId: {name: report_id, in: path, required: true, schema: {type: integer, minimum: 1}}
    PatientSession: {name: X-Patient-Session, in: header, required: true, schema: {type: string}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 8, maxLength: 128}}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 50, default: 20}}
    Offset: {name: offset, in: query, schema: {type: integer, minimum: 0, default: 0}}
  requestBodies:
    Channel:
      required: true
      content: {application/json: {schema: {type: object, properties: {channel: {type: string, enum: [SMS, WHATSAPP, EMAIL]}, language: {type: string, default: en}}}}}
  schemas:
...
    Success:
      type: object
      required: [success, data, request_id]
      properties: {success: {type: boolean, enum: [true]}, data: {type: object}, request_id: {type: string}}
    Error:
      type: object
      required: [success, error, request_id]
      properties:
        success: {type: boolean, enum: [false]}
        error: {type: object, required: [code, message], properties: {code: {type: string}, message: {type: string}}}
        request_id: {type: string}
  responses:
    Success: {description: Standard success envelope, content: {application/json: {schema: {$ref: '#/components/schemas/Success'}}}}
...
    Forbidden: {description: Scope or ownership check failed, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    Conflict: {description: State, capacity or idempotency conflict, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...